Add note about ORIGIN env var for releases
Adds a note to mention the `ORIGIN` environment variable that can be
used to control which remote repo we release to. Given that the
recommened setup for our team is to work on one of our own forks, it
makes sense that some of us have `origin` set to that fork, so this note
should clarify that we can release to something like `upstream` via
`ORIGIN=upstream make bump_and_release`, for example.